third_party/WebKit/LayoutTests/http/tests/serviceworker/registration.html

<!DOCTYPE html>
<title>Service Worker: Registration</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharness-helpers.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<script>

promise_test(function(t) {
    var script = 'resources/registration-worker.js';
    var scope = 'resources/registration/normal';
    return navigator.serviceWorker.register(script, {scope: scope})
      .then(function(registration) {
          assert_true(registration instanceof ServiceWorkerRegistration,
                      'Successfully registered.');
          service_worker_unregister_and_done(t, scope);
        })
  }, 'Registering normal scope');

promise_test(function(t) {
    var script = 'resources/registration-worker.js';
    var scope = 'resources/registration/scope-with-fragment#ref';
    return navigator.serviceWorker.register(script, {scope: scope})
      .then(function(registration) {
          assert_true(
            registration instanceof ServiceWorkerRegistration,
            'Successfully registered.');
          assert_equals(
            registration.scope,
            normalizeURL('resources/registration/scope-with-fragment'),
            'A fragment should be removed from scope')
          service_worker_unregister_and_done(t, scope);
        })
  }, 'Registering scope with fragment');

promise_test(function(t) {
    var script = 'resources/registration-worker.js';
    var scope = 'resources/';
    return navigator.serviceWorker.register(script, {scope: scope})
      .then(function(registration) {
          assert_true(registration instanceof ServiceWorkerRegistration,
                      'Successfully registered.');
          service_worker_unregister_and_done(t, scope);
        })
  }, 'Registering same scope as the script directory');

promise_test(function(t) {
    var script = 'resources/registration-worker.js';
    var scope = 'resources';
    return assert_promise_rejects(
        navigator.serviceWorker.register(script, {scope: scope}),
        'SecurityError',
        'Registering same scope as the script directory without the last ' +
            'slash should fail with SecurityError.');
  }, 'Registering same scope as the script directory without the last slash');

promise_test(function(t) {
    var script = 'resources/registration-worker.js';
    var scope = 'different-directory/';
    return assert_promise_rejects(
        navigator.serviceWorker.register(script, {scope: scope}),
        'SecurityError',
        'Registration scope outside the script directory should fail ' +
            'with SecurityError.');
  }, 'Registration scope outside the script directory');

promise_test(function(t) {
    var script = 'resources/registration-worker.js';
    var scope = 'http://example.com/';
    return assert_promise_rejects(
        navigator.serviceWorker.register(script, {scope: scope}),
        'SecurityError',
        'Registration scope outside domain should fail with SecurityError.');
  }, 'Registering scope outside domain');

promise_test(function(t) {
    var script = 'http://example.com/worker.js';
    var scope = 'http://example.com/scope/';
    return assert_promise_rejects(
        navigator.serviceWorker.register(script, {scope: scope}),
        'SecurityError',
        'Registration script outside domain should fail with SecurityError.');
  }, 'Registering script outside domain');

promise_test(function(t) {
    var script = 'resources/no-such-worker.js';
    var scope = 'resources/scope/no-such-worker';
    return assert_promise_rejects(
        navigator.serviceWorker.register(script, {scope: scope}),
        'NetworkError',
        'Registration of non-existent script should fail.');
  }, 'Registering non-existent script');

promise_test(function(t) {
    var script = 'resources/invalid-chunked-encoding.php';
    var scope = 'resources/scope/invalid-chunked-encoding/';
    return assert_promise_rejects(
        navigator.serviceWorker.register(script, {scope: scope}),
        'NetworkError',
        'Registration of invalid chunked encoding script should fail.');
  }, 'Registering invalid chunked encoding script');

promise_test(function(t) {
    var script = 'resources/invalid-chunked-encoding-with-flush.php';
    var scope = 'resources/scope/invalid-chunked-encoding-with-flush/';
    return assert_promise_rejects(
        navigator.serviceWorker.register(script, {scope: scope}),
        'NetworkError',
        'Registration of invalid chunked encoding script should fail.');
  }, 'Registering invalid chunked encoding script with flush');

promise_test(function(t) {
    var script = 'resources/plain-text-worker.php';
    var scope = 'resources/scope/plain-text-worker/';
    return assert_promise_rejects(
        navigator.serviceWorker.register(script, {scope: scope}),
        'SecurityError',
        'Registration of plain text script should fail.');
  }, 'Registering script without correct MIME type');

promise_test(function(t) {
    var script = 'resources/redirect.php?Redirect=' +
                  encodeURIComponent('/resources/registration-worker.js');
    var scope = 'resources/scope/redirect/';
    return assert_promise_rejects(
        navigator.serviceWorker.register(script, {scope: scope}),
        'SecurityError',
        'Registration of redirected script should fail.');
  }, 'Registering redirected script');

promise_test(function(t) {
    var script = 'resources/malformed-worker.php?parse-error';
    var scope = 'resources/scope/parse-error';
    return assert_promise_rejects(
        navigator.serviceWorker.register(script, {scope: scope}),
        'AbortError',
        'Registration of script including parse error should fail.');
  }, 'Registering script including parse error');

promise_test(function(t) {
    var script = 'resources/malformed-worker.php?undefined-error';
    var scope = 'resources/scope/undefined-error';
    return assert_promise_rejects(
        navigator.serviceWorker.register(script, {scope: scope}),
        'AbortError',
        'Registration of script including undefined error should fail.');
  }, 'Registering script including undefined error');

promise_test(function(t) {
    var script = 'resources/malformed-worker.php?uncaught-exception';
    var scope = 'resources/scope/uncaught-exception';
    return assert_promise_rejects(
        navigator.serviceWorker.register(script, {scope: scope}),
        'AbortError',
        'Registration of script including uncaught exception should fail.');
  }, 'Registering script including uncaught exception');

promise_test(function(t) {
    var script = 'resources/malformed-worker.php?caught-exception';
    var scope = 'resources/scope/caught-exception';
    return navigator.serviceWorker.register(script, {scope: scope})
        .then(function(registration) {
            assert_true(registration instanceof ServiceWorkerRegistration,
                        'Successfully registered.');
            service_worker_unregister_and_done(t, scope);
          })
  }, 'Registering script including caught exception');

promise_test(function(t) {
    var script = 'resources/malformed-worker.php?import-malformed-script';
    var scope = 'resources/scope/import-malformed-script';
    return assert_promise_rejects(
        navigator.serviceWorker.register(script, {scope: scope}),
        'AbortError',
        'Registration of script importing malformed script should fail.');
  }, 'Registering script importing malformed script');

promise_test(function(t) {
    var script = 'resources/malformed-worker.php?import-no-such-script';
    var scope = 'resources/scope/import-no-such-script';
    return assert_promise_rejects(
        navigator.serviceWorker.register(script, {scope: scope}),
        'AbortError',
        'Registration of script importing non-existent script should fail.');
  }, 'Registering script importing non-existent script');

</script>
